Add files via upload
Welcome to the world of Harry Potter where you can escape from the god of death by an Invisible Cloak!
the logic over here is pretty simple, it captures an image after a few msecs for future purpose. when it detects a specific type of hue (ie pink over here) it replaces that pixel with the pixel from the original image captured with respect to the coordinate. 
Note: It is still 'work under progress' as I have to configure the tolerance of hue and saturation for a given color (ie pink over here)